+++ Assets/Mushakushi.YarnSpinnerUtility/Samples~/Demo/DialogueAutoAdvance.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using Mushakushi.YarnSpinnerUtility.Runtime;
 using UnityEngine;
+using Yarn;
 using Yarn.Unity;
 
 namespace Mushakushi.YarnSpinnerUtility.Samples.Demo
@@ -9,12 +10,15 @@ namespace Mushakushi.YarnSpinnerUtility.Samples.Demo
     {
         [SerializeField] private DialogueObserver dialogueObserver;
         [SerializeField] private float waitTime = 0.5f;
+        private Coroutine autoAdvanceCoroutine;
 
         private void OnEnable()
         {
             dialogueObserver.dialogueCompleted.OnEvent += HandleDialogueCompleted;
             dialogueObserver.optionsParsed.OnEvent += HandleOptionsParsed;
             dialogueObserver.optionSelected.OnEvent += HandleOptionSelected;
+            dialogueObserver.commandParsed.OnEvent += HandleCommandParsed;
+            dialogueObserver.commandHandled.OnEvent += HandleCommandHandled;
         }
 
         private void OnDisable()
@@ -22,19 +26,31 @@ namespace Mushakushi.YarnSpinnerUtility.Samples.Demo
             dialogueObserver.dialogueCompleted.OnEvent -= HandleDialogueCompleted;
             dialogueObserver.optionsParsed.OnEvent -= HandleOptionsParsed;
             dialogueObserver.optionSelected.OnEvent -= HandleOptionSelected;
+            dialogueObserver.commandParsed.OnEvent -= HandleCommandParsed;
+            dialogueObserver.commandHandled.OnEvent -= HandleCommandHandled;
         }
 
-        private void HandleOptionsParsed(DialogueOption[] _) => StopAllCoroutines();
+        private void HandleCommandHandled()
+        {
+            autoAdvanceCoroutine = StartCoroutine(_Continue());
+        }
 
-        private void HandleOptionSelected(int _) => StartCoroutine(_Continue());
+        private void HandleOptionSelected(int _)
+        {
+            autoAdvanceCoroutine = StartCoroutine(_Continue());
+        }
 
-        private void HandleDialogueCompleted() => StopAllCoroutines();
+        private void HandleCommandParsed(Command _) => StopCoroutine(autoAdvanceCoroutine);
+
+        private void HandleOptionsParsed(DialogueOption[] _) => StopCoroutine(autoAdvanceCoroutine);
+
+        private void HandleDialogueCompleted() => StopCoroutine(autoAdvanceCoroutine);
 
         private void Start()
         {
             dialogueObserver.nodeRequested.RaiseEvent("Start");
             dialogueObserver.dialogueStarted.RaiseEvent();
-            StartCoroutine(_Continue());
+            autoAdvanceCoroutine = StartCoroutine(_Continue());
         }
 
         private IEnumerator _Continue()