+++ Assets/Mushakushi.YarnSpinnerUtility/Samples~/Demo/ExampleCommandHandler.cs
@@ -1,4 +1,5 @@
 using Mushakushi.YarnSpinnerUtility.Runtime;
+using Mushakushi.YarnSpinnerUtility.Runtime.Commands;
 using UnityEngine;
 
 namespace Mushakushi.YarnSpinnerUtility.Samples.Demo
@@ -6,22 +7,19 @@ namespace Mushakushi.YarnSpinnerUtility.Samples.Demo
     public class ExampleCommandHandler: MonoBehaviour
     {
         [SerializeField] private DialogueObserver dialogueObserver;
+        [SerializeField] private YarnCommandController commandController;
 
-        private void OnEnable()
+        private void Start()
         {
-            dialogueObserver.commandParsed.OnEvent += HandleCommandParsed; 
-        }
-
-        private void OnDisable()
-        {
-            dialogueObserver.commandParsed.OnEvent -= HandleCommandParsed;
-        }
-
-        private static void HandleCommandParsed(string[] commandElements)
-        {
-            // you can handle commands here
-            // the "wait" command is not handled by default
+            // Please note that the "wait" command is not handled by default
             // because the implementation may vary depending on usage of Awaitables, UniTask, Coroutines, etc
+            commandController.AddCommandHandler<string, string, string>("this", HandleRandomDemoCommand);
+        }
+
+        private void HandleRandomDemoCommand(string a, string b, string c)
+        {
+            Debug.Log($"Wow, th{a} {c} w{b}s handled!");
+            dialogueObserver.commandHandled.RaiseEvent();
         }
     }
 }