+++ Assets/Mushakushi.YarnSpinnerUtility/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2024-08-20
+
+### Added 
+- `CommandDispatcher` and `YarnCommandController` for to improve the command handling experience.
+
+### Changed
+- Made modifications to the Demo scene to showcase new functionality. 
+- `DialogueObserver.CommandParsed` accepts a `Comnand` object rather than a string array.
+
+### Removed 
+
+### Fixed
+- Ability to start a node that doesn't exist and continue dialogue when an option was no selected.
+
 ## [1.1.0] - 2024-08-19
 
 ### Added